Keep data when images change
With -resume=y to launch_cvd (the default), data is kept between
cuttlefish sessions, but prior to this change, it was cleared when
images were updated. This meant that system upgrades and other changes
could not be tested without losing existing data. Now, we ignore the
file modification times of images and simply warn that if problems
arise, try starting with -resume=n.

Change-Id: I5af7581613d6a074672fd1b5194ebcff97b04686